+++ src/cli/commands/private-cmds/_fetch-cmd.js
@@ -1,12 +1,11 @@
 /** @flow */
 import Command from '../../command';
-import { fromBase64 } from '../../../utils';
+import { fromBase64, unpackCommand, packCommand, buildCommandMessage } from '../../../utils';
 import { fetch } from '../../../api/scope';
 import ComponentObjects from '../../../scope/component-objects';
-import { pack } from '../../cli-utils';
 
 export default class Fetch extends Command {
-  name = '_fetch <path> <ids...>';
+  name = '_fetch <path> <args>';
   private = true;
   description = 'fetch components(s) from a scope';
   alias = '';
@@ -14,11 +13,13 @@ export default class Fetch extends Command {
     ['n', 'no_dependencies', 'do not include component dependencies']
   ];
 
-  action([path, ids, ]: [string, string[], ], { no_dependencies }: any): Promise<any> {
-    return fetch(fromBase64(path), ids.map(fromBase64), no_dependencies);
+  action([path, args]: [string, string], { no_dependencies }: any): Promise<any> {
+    const { payload } = unpackCommand(args);
+    return fetch(fromBase64(path), payload, no_dependencies);
   }
 
   report(componentObjects: ComponentObjects[]): string {
-    return pack(componentObjects.map(obj => obj.toString()));
+    const components = componentObjects.map(obj => obj.toString());
+    return packCommand(buildCommandMessage(components));
   }
 }