+++ src/consumer/component/sources/source.js
@@ -9,6 +9,6 @@ export default class Source {
   }
 
   create(bit: Bit): Source { // eslint-disable-line
-    throw Error('every source must implelement a create method');
+    throw Error('every source must implement a create method');
   }
 }