+++ src/utils/build-command-message.js
@@ -1,0 +1,10 @@
+import { BIT_VERSION } from '../constants';
+
+module.exports = function buildCommandMessage(payload) {
+  return {
+    payload,
+    headers: {
+      version: BIT_VERSION
+    }
+  };
+};