+++ src/utils/pack-command.js
@@ -1,0 +1,5 @@
+import toBase64 from './to-base64';
+
+module.exports = function packCmd(obj) {
+  return toBase64(JSON.stringify(obj));
+};